% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_cleaning.R
\name{clean_reference}
\alias{clean_reference}
\title{Clean WormExp Database Reference}
\usage{
clean_reference(
  path,
  ignore_file_names = c("c_elegans.WS283.geneIDs.txt", "reference.txt")
)
}
\arguments{
\item{path}{\code{string} Name of path where WormExp category files and
WormExp_info are saved.}

\item{ignore_file_names}{\code{character} Vector of file names to ignore.}
}
\value{
\code{NULL} (invisibly; function called for its side effect)
}
\description{
Clean WormExp Database Reference
}
\examples{
  \dontrun{
  clean_reference(
    path = "WormExp/GitHub/05_QualityManagement/WormSource_v2.0/"
    )
  }
}
\keyword{internal}
