% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_cleaning.R
\name{clean_results_table}
\alias{clean_results_table}
\title{Clean WormExp Database ResultsTable}
\usage{
clean_results_table(input_file, output_path)
}
\arguments{
\item{input_file}{\code{string} Path to WormExp database output file. Must end in
\code{".txt"}.}

\item{output_path}{\code{string} Name of path to store output.}
}
\value{
\code{NULL} (invisibly; function called for its side effect)
}
\description{
This function cleans the WormExp database output and saves it back as .txt
file.
}
\examples{
  \dontrun{
  clean_results_table(
    input_file = "WormExp/results/Test_ResultsTable.txt",
    output_path = "WormExp/results"
    )
  }
}
