% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text_files.R
\name{clean_text_files}
\alias{clean_text_files}
\title{Remove Duplicates From WormExp Text Files}
\usage{
clean_text_files(path, output_path)
}
\arguments{
\item{path}{\code{string} Path where to find the WormExp database output.}

\item{output_path}{\code{string} Path to write the processed files to.}
}
\value{
\code{NULL} (invisibly; function called for its side effect)
}
\description{
This function cleans the WormExp ResultsTable and saves it back as .txt file
}
\examples{
  \dontrun{
  clean_text_files(
    path = "WormExp/GitHub/05_QualityManagement/WormSource_v1.5/",
    output_path = "WormExp/GitHub/05_QualityManagement/WormSource_v1.5/cleaned/"
    )
  }
}
\keyword{internal}
