% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{compare_results}
\alias{compare_results}
\title{Compare WormExp Query Results}
\usage{
compare_results(
  input_path,
  output_path,
  top = 30,
  category = NULL,
  width = 1200,
  height = 1200
)
}
\arguments{
\item{input_path}{\code{string} Name of path containing cleaned WormExp query
output data files.}

\item{output_path}{\code{string} Name of path to store output.}

\item{top}{\code{integer} Number of top hits to consider.}

\item{category}{\code{string} Name of the category to restrict the comparison to.
\code{NULL} for no restriction.}

\item{width}{\code{integer} Width of png output (in pixels).}

\item{height}{\code{integer} Height of png output (in pixels).}
}
\value{
\code{NULL} (invisibly; function called for its side effect)
}
\description{
This function takes the columns Category, Term and OverlappedID and presents
a heatmap that shows the percentage of genes overlapping within all top hits.
In addition, the user can provide the name of a specific category to which
the comparison should be restricted.
}
\details{
Use \code{\link[=clean_results_table]{clean_results_table()}} to prepare your WormExp query output for this
function. Resulting heatmaps are \emph{NOT SYMMETRIC} and have to be read
column-wise.
}
\examples{
  \dontrun{
  compare_results(
    input_path = "WormExp/results",
    output_path = "WormExp/figures"
    )
  }
}
\seealso{
\code{\link[=clean_results_table]{clean_results_table()}}
}
