% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{import}
\alias{import}
\title{Import Data From Files}
\usage{
import(path, pattern, ignore_pattern = NULL, header)
}
\arguments{
\item{path}{\code{string} Name of path with data files.}

\item{pattern}{\code{string} Regular expression for selecting data files.}

\item{ignore_pattern}{Regular expression for data files to be ignored.}

\item{header}{\code{logical} Use headers of tabular data?}
}
\value{
A \code{list}.
}
\description{
Returns a list of data from all files in a given folder.
}
\keyword{internal}
