% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{is_integer}
\alias{is_integer}
\title{Test For Integer Value}
\usage{
is_integer(x)
}
\arguments{
\item{x}{\code{numeric} or \code{integer} scalar or vector to test.}
}
\value{
\code{TRUE} if \code{x} has an integer value, \code{FALSE} otherwise.
}
\description{
Checks whether a given number has an integer value.
}
\keyword{internal}
