% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{is_valid_path}
\alias{is_valid_path}
\title{Test Path Validity}
\usage{
is_valid_path(path)
}
\arguments{
\item{path}{\code{string} Name of path to test.}
}
\value{
\code{TRUE} if \code{path} is a valid (existing) path name, \code{FALSE} otherwise.
}
\description{
Checks whether a given path name actually exists.
}
\keyword{internal}
