% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{trim}
\alias{trim}
\title{Trim Leading And Trailing Spaces}
\usage{
trim(x)
}
\arguments{
\item{x}{\code{character} Vector, possibly containing leading and trailing spaces.}
}
\value{
A \code{character} vector.
}
\description{
Trim Leading And Trailing Spaces
}
\keyword{internal}
