% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WormExpR-package.R
\docType{package}
\name{WormExpR-package}
\alias{WormExpR}
\alias{WormExpR-package}
\title{WormExpR: WormExp Database Query Cleaning and Visualization}
\description{
Provides functions for cleaning and visualizing
\href{https://wormexp.zoologie.uni-kiel.de/wormexp/}{WormExp} query results,
internal functions for database maintenance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/omics2view/WormExpR}
  \item Report bugs at \url{https://github.com/omics2view/WormExpR/issues}
}

}
\author{
\strong{Maintainer}: Hinrich Schulenburg \email{hschulenburg@zoologie.uni-kiel.de}

Authors:
\itemize{
  \item Evolutionary Ecology and Genetics, University Kiel, Germany
  \item Jennifer Neumaier
}

Other contributors:
\itemize{
  \item omics2view.consulting GbR, Kiel, Germany [compiler, contractor]
}

}
